+++ src/main/java/prismatic/shards/stellarity/registry/StellarityFeatures.java
@@ -3,17 +3,19 @@ package prismatic.shards.stellarity.registry;
 import net.minecraft.core.Registry;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.level.levelgen.feature.Feature;
+import net.minecraft.world.level.levelgen.feature.configurations.SimpleBlockConfiguration;
 import prismatic.shards.stellarity.Stellarity;
-import prismatic.shards.stellarity.registry.feature.DungeonFeature;
-import prismatic.shards.stellarity.registry.feature.FreezeWaterFeature;
-import prismatic.shards.stellarity.registry.feature.SpikeFeature;
+import prismatic.shards.stellarity.registry.feature.*;
 import prismatic.shards.stellarity.registry.feature.configuration.DungeonFeatureConfiguration;
+import prismatic.shards.stellarity.registry.feature.configuration.FeatureSequenceConfiguration;
 import prismatic.shards.stellarity.registry.feature.configuration.SpikeFeatureConfiguration;
 
 public interface StellarityFeatures {
 	DungeonFeature DUNGEON = register("dungeon", new DungeonFeature(DungeonFeatureConfiguration.CODEC));
 	FreezeWaterFeature FREEZE_WATER = register("freeze_water", new FreezeWaterFeature());
 	SpikeFeature SPIKE = register("spike", new SpikeFeature(SpikeFeatureConfiguration.CODEC));
+	DragonEggFeature DRAGON_EGG = register("dragon_egg", new DragonEggFeature(SimpleBlockConfiguration.CODEC));
+	FeatureSequenceFeature FEATURE_SEQUENCE = register("feature_sequence", new FeatureSequenceFeature(FeatureSequenceConfiguration.CODEC));
 
 	static <T extends Feature<?>> T register(String id, T feature) {
 		return Registry.register(BuiltInRegistries.FEATURE, Stellarity.id(id), feature);